requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==1.1.2
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
openpyxl==3.0.7
pandas==1.3.2
pymssql==2.2.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==2.0.1
